render terminal with formatting
this was surprisingly painful, but there's probably a better way to do
it. this feels pretty inefficient.

it's pretty funny that we parse ANSI escape sequences and simulate them
in a virtual terminal just to painstakingly reconstruct the same output
to the original terminal, but it's kind of necessary to not obliterate
the terminal when running fancy commands in parallel