always sort vertexes chronologically
this feels a million times more intuitive; right now they're sorted in
the order that they're started, which in recursive cases will end up
with the outer vertexes first, meaning their final logs will be ahead of
the things that they ran.

this way you can just always watch the bottom of the screen for the most
recent activity.