vertexes can be members of multiple groups
Recorder will now keep track of groups associated to vertex digests,
globally to the recorder. Whenever an update is recorded, all vertex
groups will be expanded to the union of all groups ever seen by the
recorder.

This goes for both the Go API and direct calls to Record (as when
converting SolveStatuses from Buildkit).