rpc: guard grpc status stream send w/ mutex
gRPC has a fun quirk where if you call SendMsg on the same stream in
parallel from multiple goroutines, you can easily hit a deadlock in some
internal grpc writeQuota thing:
https://github.com/grpc/grpc-go/issues/5393#issuecomment-1148851417

Was hitting some random deadlocks in dagger and saw from SIGQUIT stacks
that we were blocked on that writeQuota call. Wrapping the WriteStatus
implementation in a mutex fixed it.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>