new statusbar, used for all modes
no more Big Ugly Bar in focus mode; now it just shows the count for each
status instead, matching the icons used in focus mode's tree UI.

a better mental model is that now there are two UIs, tree and graph, and
focus mode always uses the tree UI. there's probably a better way to
refactor this.

the full run duration is now displayed too.

the help UI is shimmied onto the end of the last status info to stay
out of the way and optimize screen space, but I'll admit it's a bit
jank.