reappropriate tui param with interactive
this param was always true in practice. it may have originally been how
to toggle console mode, but console mode works entirely differently now.
it seems more useful to treat this as an interactive on/off, so it can
be disabled for things that require interaction, like REPLs or shells.

along the way:

* go back to viewport-based rendering
* render chrome in the same loop as the regular render loop, so that we
  can take its height into account when rendering. this fixed a couple
  TODOs for hardcoding "- 2".
* use persistent buffer for the render loop, likely major perf low
  hanging fruit