major refactor to more conventional API
The async RenderLoop API was pretty complicated to use. Everything gets
a lot simpler when you pass a callback instead, because you can just
wrap it in a defer to restore the terminal to "cooked" mode, integrate
the callback lifecycle into the Model, and just print the final render
after the program exits.